requirements.txt
google-adk==1.28.0
python-dotenv
requests